requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
aiofiles==24.1.0

# Configuration
pydantic-settings==2.5.2
python-dotenv==1.0.1

# Redis (job state management)
redis[hiredis]==5.2.0

# Celery (optional - for production scaling)
celery[redis]==5.4.0

# S3-compatible storage
boto3==1.35.0
aioboto3==13.1.1

# OpenAI (translation)
openai==1.51.0

# ElevenLabs (TTS)
elevenlabs==1.9.0

# Speaker diarization
pyannote.audio==3.3.2
torch==2.4.0
torchaudio==2.4.0

# Speech-to-text (faster Whisper)
faster-whisper==1.0.3

# Audio processing
pydub==0.25.1
librosa==0.10.2
soundfile==0.12.1
numpy==1.26.4

# HTTP client
httpx==0.27.2
tenacity==9.0.0

# Utilities
python-jose[cryptography]==3.3.0
structlog==24.4.0
